im/requirements.txt
pymongo
flask
flask-cors
pillow